chore: add useRetry
