+++ packages/maskbook/src/plugins/Snapshot/UI/NetworkFail.tsx
@@ -2,14 +2,27 @@ import { Component } from 'react'
 import { SnapshotCard } from './SnapshotCard'
 import { Typography, Button } from '@material-ui/core'
 
-export class NetworkFail extends Component<{ title: string; retry: () => void }> {
+export class NetworkFail extends Component<{ title: string; retry: () => void; isFullPluginDown?: boolean }> {
     static getDerivedStateFromError(error: unknown) {
         return { error }
     }
     state: { error: Error | null } = { error: null }
     render() {
         if (this.state.error) {
-            return (
+            return this.props.isFullPluginDown ? (
+                <>
+                    <Typography>Loading fail due to Snapshot api service breakdown.</Typography>
+                    <Button
+                        style={{ width: 100, marginTop: 16 }}
+                        variant="outlined"
+                        onClick={() => {
+                            this.setState({ error: null })
+                            this.props.retry()
+                        }}>
+                        Retry
+                    </Button>
+                </>
+            ) : (
                 <SnapshotCard title={this.props.title}>
                     <Typography>Loading fail due to Snapshot api service breakdown.</Typography>
                     <Button
