+++ packages/maskbook/src/plugins/Snapshot/UI/PostInspector.tsx
@@ -1,6 +1,8 @@
 import { SnapshotContext } from '../context'
 import { useProposalIdentifier } from '../hooks/useProposalIdentifier'
 import { Snapshot } from './Snapshot'
+import { NetworkFail } from './NetworkFail'
+import { useRetry } from '../hooks/useRetry'
 
 export interface PostInspectorProps {
     url: string
@@ -8,9 +10,12 @@ export interface PostInspectorProps {
 
 export function PostInspector(props: PostInspectorProps) {
     const identifier = useProposalIdentifier(props.url)
+    const retry = useRetry()
     return (
         <SnapshotContext.Provider value={identifier}>
-            <Snapshot />
+            <NetworkFail title="" isFullPluginDown={true} retry={retry}>
+                <Snapshot />
+            </NetworkFail>
         </SnapshotContext.Provider>
     )
 }
